# Prefer to take AOT method from the SC queue faster