@vnkozlov 2024-09-06
Currently we do that in `compare_task`:
https://github.com/openjdk/leyden/blob/premain/src/hotspot/share/compiler/compilationPolicy.cpp#L1185

I was very conservative to immediately selecting SCC task because we put these task on normal compilation queues:
https://github.com/veresov/jdk/commit/76157f6e77346ef0b2c47497231210f1fd84b446

Now we have separate queues for these tasks. I thought I changed code how task is selected for these queues (using FIFO).