@shipilev 2024-09-07
> Currently we do that in `compare_task`:
> https://github.com/openjdk/leyden/blob/premain/src/hotspot/share/compiler/compilationPolicy.cpp#L1185

Aha! That is not efficient, since we are still walking the entire queue. I measured that task selection on 1K queue takes ~25us, which means for 2K elements we loiter in this code for 50ms, delaying the AOT loading. This explains why this patch has such an effect.

I dropped the SCC checks from `compare_task` and replaced them with asserts, since we are not expecting SCC tasks in that comparator anymore. No need to check SCC for normal compilation queues, which might delay compilation a bit.